+++ src/component/header.js
@@ -2,7 +2,14 @@ import React, { Component } from 'react'
 import {Link} from 'react-router-dom'
 
 class Header extends Component {
+
+    state= {
+        quantity:0,
+    }
+
     render () {
+        console.log(this.state);
+       let noOfItem = this.handleLanguage;
         return (
             <header className="header1">
                 <div className="container-menu-header">
@@ -51,7 +58,7 @@ class Header extends Component {
 
                         <div className="header-wrapicon2">
                             <img src="images/icons/icon-header-02.png" className="header-icon1 js-show-header-dropdown" alt="ICON"/>
-                            <span className="header-icons-noti">0</span>
+                            <span className="header-icons-noti">{noOfItem}</span>
                         </div>
                     </div>
