+++ src/component/main.css
@@ -40,7 +40,7 @@ Input, textarea : 14px/1.6 '', Arial, sans-serif;
 ///////////////////////////////////////////////////////////
 */
 
-@font-face {
+/* @font-face {
   font-family: Montserrat-Regular;
   src: url('../fonts/montserrat/Montserrat-Regular.ttf'); 
 }
@@ -78,7 +78,7 @@ Input, textarea : 14px/1.6 '', Arial, sans-serif;
 @font-face {
   font-family: Poppins-Black;
   src: url('../fonts/poppins/Poppins-Black.ttf'); 
-}
+} */
 
 /*[ RESTYLE TAG ]
 ///////////////////////////////////////////////////////////
